pyproject.toml
[project]
name = "django-edtech-platforms"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=6.0.4",
    "django-ckeditor>=6.7.3",
    "django-tailwind[cookiecutter,honcho,reload]>=4.4.2",
    "python-decouple>=3.8",
]
